fix: update all the URLs ahead of full transfer
BREAKING CHANGE: The ownership of this package has officially been transferred to https://github.com/favna.
While Favna has been maintaining this package for a few years already and this won't affect you in any major ways,
package ownership transfers have often been an icky topic in the NPM ecosystem. For this reason the choice was made
to make this a new major versions so a change of ownership would not land in your caret scoped dependencies.